Corrections do GlobalHeader. Still bugs due to render of IconSettings
