+++ components/global-header/global-header-dropdown/global-header-dropdown.jsx
@@ -10,16 +10,16 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 // ### Dropdown
-import MenuDropdown from '../../menu-dropdown';
+import MenuDropdown from '../../menu-dropdown/menu-dropdown';
 import GlobalHeaderTrigger from '../private/dropdown-trigger';
 
 // ## Constants
-import { GLOBAL_HEADER_TOOL } from '../../../utilities/constants';
+import { GLOBAL_HEADER_DROPDOWN } from '../../../utilities/constants';
 
 /**
  * This component is an implementation of `MenuDropdown` with a custom trigger. All the properties listed below are provided to the `MenuDropdown` component. Any additional properties are provided to the Custom Trigger (that is the `Button` or `li` tag).
  */
-const GlobalHeaderDropdown = (props) => {
+function GlobalHeaderDropdown(props) {
 	const { globalAction, iconVariant, ...rest } = props;
 	let iconVariantOverride;
 
@@ -41,7 +41,7 @@ const GlobalHeaderDropdown = (props) => {
 // ### Display Name
 // Always use the canonical component name (set in the core) as the React
 // display name.
-GlobalHeaderDropdown.displayName = GLOBAL_HEADER_TOOL;
+GlobalHeaderDropdown.displayName = GLOBAL_HEADER_DROPDOWN;
 
 // ### Prop Types
 GlobalHeaderDropdown.propTypes = {
@@ -113,7 +113,6 @@ GlobalHeaderDropdown.propTypes = {
 GlobalHeaderDropdown.defaultProps = {
 	align: 'right',
 	buttonVariant: 'icon',
-	iconVariant: 'global-header',
 	nubbinPosition: 'top right',
 };
