+++ components/modal/presets/0-default.jsx
@@ -1,0 +1,13 @@
+/* eslint-disable filenames/match-regex */
+import * as React from 'react';
+import Modal from '../modal.jsx';
+
+export default (
+	<Modal uxpId="0">
+		Dolor eiusmod sunt ex incididunt cillum quis nostrud velit duis sit officia.
+		Lorem aliqua enim laboris do dolor eiusmod officia. Mollit incididunt nisi
+		consectetur esse laborum eiusmod pariatur proident. Eiusmod et adipisicing
+		culpa deserunt nostrud ad veniam nulla aute est. Labore esse esse cupidatat
+		amet velit id elit consequat minim ullamco mollit enim excepteur ea.
+	</Modal>
+);
