+++ utilities/constants.js
@@ -1,107 +1,107 @@
 // ## Components
 export const ACCORDION = 'Accordion';
 export const ACCORDION_PANEL = 'AccordionPanel';
-export const ALERT = 'SLDSAlert';
-export const ALERT_CONTAINER = 'SLDSAlertContainer';
+export const ALERT = 'Alert';
+export const ALERT_CONTAINER = 'AlertContainer';
 export const APP_LAUNCHER = 'AppLauncher';
 export const APP_LAUNCHER_SECTION = 'AppLauncherSection';
 export const APP_LAUNCHER_TILE = 'AppLauncherTile';
 export const APP_LAUNCHER_HEADER = 'AppLauncherHeader';
-export const AVATAR = 'SLDSAvatar';
-export const BRAND_BAND = 'SLDSBrandBand';
+export const AVATAR = 'Avatar';
+export const BRAND_BAND = 'BrandBand';
 
 /* Remove at next breaking change */
-export const BREAD_CRUMB = 'SLDSBreadcrumb';
+export const BREAD_CRUMB = 'Breadcrumb';
 
-export const BREADCRUMB = 'SLDSBreadcrumb';
+export const BREADCRUMB = 'Breadcrumb';
 export const BUTTON = 'Button';
 export const BUTTON_GROUP = 'ButtonGroup';
 export const BUTTON_STATEFUL = 'ButtonStateful';
-export const BUTTON_ICON = 'SLDSButtonIcon'; // a non-interctive icon wrapper for SLDSUtilityIcon
-export const CARD = 'SLDSCard';
-export const CARD_BODY = 'SLDSCardBody';
-export const CARD_EMPTY = 'SLDSCardEmpty';
-export const CARD_FILTER = 'SLDSCardFilter';
-export const CARD_FOOTER = 'SLDSCardFooter';
-export const CARD_HEADER = 'SLDSCardHeader';
-export const COLOR_PICKER = 'SLDSColorPicker';
-export const COMBOBOX = 'SLDSCombobox';
-export const DATA_TABLE = 'SLDSDataTable';
-export const DATA_TABLE_CELL = 'SLDSDataTableCell';
-export const DATA_TABLE_COLUMN = 'SLDSDataTableColumn';
-export const DATA_TABLE_HEAD = 'SLDSDataTableHead';
-export const DATA_TABLE_HEADER_CELL = 'SLDSDataTableHeaderCell';
-export const DATA_TABLE_ROW_ACTIONS = 'SLDSDataTableRowActions';
-export const DATA_TABLE_ROW = 'SLDSDataTableRow';
-export const DATE_PICKER = 'SLDSDatePicker';
-export const DIALOG = 'SLDSDialog';
-export const FILTER = 'SLDSFilter';
-export const CHECKBOX = 'SLDSCheckbox';
-export const FORMS_INLINE_EDIT = 'SLDSInlineEdit';
-export const INPUT = 'SLDSInput';
-export const TEXTAREA = 'SLDSTextarea';
-export const SEARCH = 'SLDSSearch';
-export const GLOBAL_HEADER = 'SLDSGlobalHeader';
-export const GLOBAL_HEADER_PROFILE = 'SLDSGlobalHeaderProfile';
-export const GLOBAL_HEADER_SEARCH = 'SLDSGlobalHeaderSearch';
-export const GLOBAL_HEADER_TOOL = 'SLDSGlobalHeaderTool';
-export const GLOBAL_NAVIGATION_BAR = 'SLDSGlobalNavigationBar';
-export const GLOBAL_NAVIGATION_BAR_BUTTON = 'SLDSGlobalNavigationBarButton';
-export const GLOBAL_NAVIGATION_BAR_DROPDOWN = 'SLDSGlobalNavigationBarDropdown';
-export const GLOBAL_NAVIGATION_BAR_LABEL = 'SLDSGlobalNavigationBarLabel';
-export const GLOBAL_NAVIGATION_BAR_LINK = 'SLDSGlobalNavigationBarLink';
-export const GLOBAL_NAVIGATION_BAR_REGION = 'SLDSGlobalNavigationBarRegion';
+export const BUTTON_ICON = 'ButtonIcon'; // a non-interctive icon wrapper for UtilityIcon
+export const CARD = 'Card';
+export const CARD_BODY = 'CardBody';
+export const CARD_EMPTY = 'CardEmpty';
+export const CARD_FILTER = 'CardFilter';
+export const CARD_FOOTER = 'CardFooter';
+export const CARD_HEADER = 'CardHeader';
+export const COLOR_PICKER = 'ColorPicker';
+export const COMBOBOX = 'Combobox';
+export const DATA_TABLE = 'DataTable';
+export const DATA_TABLE_CELL = 'DataTableCell';
+export const DATA_TABLE_COLUMN = 'DataTableColumn';
+export const DATA_TABLE_HEAD = 'DataTableHead';
+export const DATA_TABLE_HEADER_CELL = 'DataTableHeaderCell';
+export const DATA_TABLE_ROW_ACTIONS = 'DataTableRowActions';
+export const DATA_TABLE_ROW = 'DataTableRow';
+export const DATE_PICKER = 'DatePicker';
+export const DIALOG = 'Dialog';
+export const FILTER = 'Filter';
+export const CHECKBOX = 'Checkbox';
+export const FORMS_INLINE_EDIT = 'InlineEdit';
+export const INPUT = 'Input';
+export const TEXTAREA = 'Textarea';
+export const SEARCH = 'Search';
+export const GLOBAL_HEADER = 'GlobalHeader';
+export const GLOBAL_HEADER_PROFILE = 'GlobalHeaderProfile';
+export const GLOBAL_HEADER_SEARCH = 'GlobalHeaderSearch';
+export const GLOBAL_HEADER_DROPDOWN = 'GlobalHeaderDropdown';
+export const GLOBAL_NAVIGATION_BAR = 'GlobalNavigationBar';
+export const GLOBAL_NAVIGATION_BAR_BUTTON = 'GlobalNavigationBarButton';
+export const GLOBAL_NAVIGATION_BAR_DROPDOWN = 'GlobalNavigationBarDropdown';
+export const GLOBAL_NAVIGATION_BAR_LABEL = 'GlobalNavigationBarLabel';
+export const GLOBAL_NAVIGATION_BAR_LINK = 'GlobalNavigationBarLink';
+export const GLOBAL_NAVIGATION_BAR_REGION = 'GlobalNavigationBarRegion';
 export const GLOBAL_NAVIGATION_BAR_APP_LAUNCHER =
-	'SLDSGlobalNavigationBarAppLauncher';
-export const GRID = 'SLDSGrid';
-export const HIGHLIGHTER = 'SLDSHighlighter';
+	'GlobalNavigationBarAppLauncher';
+export const GRID = 'Grid';
+export const HIGHLIGHTER = 'Highlighter';
 export const ICON = 'Icon';
-export const ICON_SETTINGS = 'SLDSIconSettings';
-export const ICON_INPUT = 'SLDSIconInput';
-export const ILLUSTRATION = 'SLDSIllustration';
-export const LIST = 'SLDSList';
-export const LIST_ITEM = 'SLDSListItem';
-export const LIST_ITEM_LABEL = 'SLDSListItemLabel';
-export const LISTBOX_OF_PILL_OPTIONS = 'SLDSListboxOfPillOptions';
-export const LOOKUP = 'SLDSLookup';
-export const MEDIA_OBJECT = 'SLDSMediaObject';
+export const ICON_SETTINGS = 'IconSettings';
+export const ICON_INPUT = 'IconInput';
+export const ILLUSTRATION = 'Illustration';
+export const LIST = 'List';
+export const LIST_ITEM = 'ListItem';
+export const LIST_ITEM_LABEL = 'ListItemLabel';
+export const LISTBOX_OF_PILL_OPTIONS = 'ListboxOfPillOptions';
+export const LOOKUP = 'Lookup';
+export const MEDIA_OBJECT = 'MediaObject';
 export const MENU_DROPDOWN = 'MenuDropdown';
-export const MENU_DROPDOWN_TRIGGER = 'SLDSMenuDropdownTrigger';
-export const MENU_PICKLIST = 'SLDSMenuPicklist';
+export const MENU_DROPDOWN_TRIGGER = 'MenuDropdownTrigger';
+export const MENU_PICKLIST = 'MenuPicklist';
 export const MODAL = 'Modal';
-export const NOTIFICATION = 'SLDSNotification';
-export const PANEL = 'SLDSPanel';
-export const PANEL_FILTER_GROUP = 'SLDSFilterGroup';
-export const PANEL_FILTER_LIST = 'SLDSFilterList';
-export const PANEL_FILTER_LIST_HEADING = 'SLDSPanelFilterListHeading';
-export const PAGE_HEADER = 'SLDSPageHeader';
+export const NOTIFICATION = 'Notification';
+export const PANEL = 'Panel';
+export const PANEL_FILTER_GROUP = 'FilterGroup';
+export const PANEL_FILTER_LIST = 'FilterList';
+export const PANEL_FILTER_LIST_HEADING = 'PanelFilterListHeading';
+export const PAGE_HEADER = 'PageHeader';
 export const PILL = 'Pill';
-export const PILL_CONTAINER = 'SLDSPillContainer';
+export const PILL_CONTAINER = 'PillContainer';
 export const POPOVER = 'Popover';
 export const POPOVER_TOOLTIP = 'PopoverTooltip';
-export const PROGRESS_INDICATOR = 'SLDSProgressIndicator';
-export const PROGRESS_INDICATOR_PROGRESS = 'SLDSProgressIndicatorProgress';
-export const PROGRESS_INDICATOR_STEP = 'SLDSProgressIndicatorStep';
-export const PROGRESS_RING = 'SLDSProgressRing';
-export const SLIDER = 'SLDSSlider';
-export const SPINNER = 'SLDSSpinner';
-export const SPLIT_VIEW = 'SLDSSplitView';
-export const SPLIT_VIEW_HEADER = 'SLDSSplitViewHeader';
-export const SPLIT_VIEW_LISTBOX = 'SLDSSplitViewListbox';
-export const RADIO = 'SLDSRadio';
+export const PROGRESS_INDICATOR = 'ProgressIndicator';
+export const PROGRESS_INDICATOR_PROGRESS = 'ProgressIndicatorProgress';
+export const PROGRESS_INDICATOR_STEP = 'ProgressIndicatorStep';
+export const PROGRESS_RING = 'ProgressRing';
+export const SLIDER = 'Slider';
+export const SPINNER = 'Spinner';
+export const SPLIT_VIEW = 'SplitView';
+export const SPLIT_VIEW_HEADER = 'SplitViewHeader';
+export const SPLIT_VIEW_LISTBOX = 'SplitViewListbox';
+export const RADIO = 'Radio';
 export const RADIO_GROUP = 'RadioGroup';
 export const RADIO_BUTTON_GROUP = 'RadioButtonGroup';
-export const TAB = 'SLDSTab';
-export const TAB_ITEM = 'SLDSTabItem';
-export const TAB_PANEL = 'SLDSTabPanel';
+export const TAB = 'Tab';
+export const TAB_ITEM = 'TabItem';
+export const TAB_PANEL = 'TabPanel';
 export const TABS = 'Tabs';
-export const TABS_LIST = 'SLDSTabsList';
+export const TABS_LIST = 'TabsList';
 export const TABS_PANEL = 'TabsPanel';
-export const TIME_PICKER = 'SLDSTimepicker';
-export const TOAST = 'SLDSToast';
-export const TOAST_CONTAINER = 'SLDSToastContainer';
+export const TIME_PICKER = 'Timepicker';
+export const TOAST = 'Toast';
+export const TOAST_CONTAINER = 'ToastContainer';
 export const TREE = 'Tree';
 export const TREE_BRANCH = 'TreeBranch';
 export const TREE_ITEM = 'TreeItem';
-export const VERTICAL_NAVIGATION = 'SLDSVerticalNavigation';
-export const VERTICAL_NAVIGATION_ITEM = 'SLDSVerticalNavigationItem';
+export const VERTICAL_NAVIGATION = 'VerticalNavigation';
+export const VERTICAL_NAVIGATION_ITEM = 'VerticalNavigationItem';
